Compiler semantic pass
- Rename `CScriptMacros` to `CScriptCompiler`
- Make authoritative over unlocalization data
- Simplify `CScriptMacro` implementation
