+++ gamedata/scripts/macro_lua.script
@@ -1,6 +1,6 @@
-function lua(src, namespace_name, unlocalizer)
+function expand(src, namespace_name, unlocalizer)
    return string.format(
-      [[
+      [=[
 local function f()
    local function script_name()
       return "%s"
@@ -10,7 +10,7 @@ local function f()
 end
 
 package.loaded["%s"] = f()
-      ]],
+      ]=],
       namespace_name,
       src,
       namespace_name
