+++ src/xrGame/console_commands.cpp
@@ -1713,7 +1713,7 @@ public:
 			string4096 S;
 			shared_str m_script_name = "console command";
 			xr_sprintf(S, "%s\n", args);
-			bool loaded = ai().script_engine().load_buffer(ai().script_engine().lua(), NULL, S, xr_strlen(S), *m_script_name);
+			bool loaded = ai().script_engine().load_buffer(ai().script_engine().lua(), S, xr_strlen(S), *m_script_name);
 			if (loaded)
 			{
 				int l_iErrorCode = lua_pcall(ai().script_engine().lua(), 0, 0, 0);
