+++ src/xrGame/vs2022/xrGame.vcxproj.filters
@@ -2498,6 +2498,7 @@
       <UniqueIdentifier>{c14ffa0c-947d-49b9-8e01-269569b48371}</UniqueIdentifier>
     </Filter>
     <Filter Include="AI\AScript\ScriptDialect">
+    <Filter Include="AI\AScript\ScriptCompiler">
       <UniqueIdentifier>{53abd459-f211-4178-b3c6-2be2afe329ba}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
@@ -7409,16 +7410,16 @@
     </ClInclude>
     <ClInclude Include="..\..\xrServerEntities\script_dialect.h">
     <ClInclude Include="..\..\xrServerEntities\script_macro.h">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\xrServerEntities\script_macros.h">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+    <ClInclude Include="..\..\xrServerEntities\script_compiler.h">
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClInclude>
     <ClInclude Include="..\..\xrServerEntities\lua_macros.h">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClInclude>
     <ClInclude Include="..\..\xrServerEntities\script_macro_wua.h">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -11129,16 +11130,16 @@
     </ClCompile>
     <ClCompile Include="..\..\xrServerEntities\script_dialect_lisp_macro.cpp">
     <ClCompile Include="..\..\xrServerEntities\script_macro.cpp">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\xrServerEntities\script_macros.cpp">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+    <ClCompile Include="..\..\xrServerEntities\script_compiler.cpp">
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClCompile>
     <ClCompile Include="..\..\xrServerEntities\script_macro_wua.cpp">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClCompile>
     <ClCompile Include="..\..\xrServerEntities\script_macro_script.cpp">
-      <Filter>AI\AScript\ScriptDialect</Filter>
+      <Filter>AI\AScript\ScriptCompiler</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
