+++ src/xrServerEntities/script_engine.cpp
@@ -13,6 +13,7 @@
 #include "script_process.h"
 #include "../build_config_defines.h"
 #include "script_storage.h"
+#include "script_compiler.h"
 #include <unordered_map>
 #include <set>
 
@@ -354,8 +355,6 @@ void CScriptEngine::setup_auto_load()
 }
 
 extern void export_classes(lua_State* L);
-extern xr_unordered_map<std::string, std::set<std::string>> unlocalizers;
-extern bool unlocalizerPassed;
 
 void CScriptEngine::init()
 {
@@ -402,8 +401,7 @@ void CScriptEngine::init()
 #endif // #ifndef USE_LUA_STUDIO
 	//	lua_sethook							(lua(), lua_hook_call,	LUA_MASKLINE|LUA_MASKCALL|LUA_MASKRET,	0);
 
-	unlocalizers.clear();
-	unlocalizerPassed = false;
+	ScriptCompiler().load_unlocalizers();
 	bool save = m_reload_modules;
 	m_reload_modules = true;
 	process_file_if_exists("_G", false);
