+++ src/xrServerEntities/script_macro_wua.h
@@ -5,6 +5,5 @@
 class CWuaMacro : public CScriptMacro
 {
 public:
-    virtual std::string unlocalize(const std::string& src, LPCSTR caNameSpaceName, Unlocalizer& unlocalizer) const override;
-    virtual std::string lift(const std::string& src, LPCSTR caNameSpaceName) const override;
+    virtual std::string lift(std::string src, LPCSTR caNameSpaceName) const override;
 };
