+++ src/xrServerEntities/script_thread.cpp
@@ -14,7 +14,7 @@
 };*/
 //-AVO
 #include "script_engine.h"
-#include "script_macros.h"
+#include "script_compiler.h"
 #include "script_thread.h"
 #include "ai_space.h"
 
@@ -57,7 +57,7 @@ CScriptThread::CScriptThread(LPCSTR caBuffer, bool do_string, bool reload)
 		{
 			m_script_name = "console command";
 			S += caBuffer;
-			S = ScriptMacros().lift(S, *m_script_name);
+			S = ScriptCompiler().lift(S, *m_script_name);
 			S = "function " + std::string(main_function) + "()\n" + S + "\nend";
 			int l_iErrorCode = luaL_loadbuffer(ai().script_engine().lua(), S.c_str(), S.length(), "@console_command");
 			if (!l_iErrorCode)
@@ -114,7 +114,7 @@ CScriptThread::CScriptThread(LPCSTR caBuffer, bool do_string, bool reload)
 		else
 			S = std::string(main_function) + "()";
 
-		if (!ai().script_engine().load_buffer(lua(), NULL, S.c_str(), S.length(), "@_thread_main"))
+		if (!ai().script_engine().load_buffer(lua(), S.c_str(), S.length(), "@_thread_main"))
 			return;
 
 		m_active = true;
